requirements.txt
slack_bolt
python-dotenv
openpyxl
pandas
